+++ doc/conf.py
@@ -41,7 +41,7 @@ MOCK_MODULES = [
     'matplotlib',
     'matplotlib.mlab', 'matplotlib.pyplot', 'matplotlib.widgets',
     'matplotlib.nxutils',
-    'PIL.Image', 'fatiando.seis._cwavefd']
+    'PIL.Image']
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = Mock()