+++ fatiando/seis/wavefd.py
@@ -3,9 +3,10 @@ Finite difference solution of the 2D wave equation for isotropic media.
 
 .. warning::
 
-    Due to the high computational demmand of these simulations, there
-    are no pure Python alternatives to the optimized Cython modules. For
-    instructions on using and installing the Cython compiled modules, see
+    Due to the high computational demmand of these simulations,
+    the pure Python time stepping functions are **very** slow!
+    I strongly recommend using the optimized Cython time stepping module.
+    For instructions on using and installing the Cython compiled modules, see
     :ref:`advanced-usage`.
 
 Simulates both elastic and acoustic waves:
@@ -180,12 +181,14 @@ The solution for P and SV waves is:
 """
 import numpy
 
-import fatiando.seis._cwavefd
 import fatiando.log
 
 log = fatiando.log.dummy('fatiando.seis.wavefd')
 
-
+try:
+    from fatiando.seis import _cwavefd as timestepper
+except ImportError:
+    from fatiando.seis import _wavefd as timestepper
 
 class MexHatSource(object):
     r"""
@@ -370,10 +373,10 @@ def elastic_sh(spacing, shape, svel, dens, deltat, iterations, sources,
 
     """
     nz, nx = shape
-    dz, dx = spacing
+    dz, dx = (float(i) for i in spacing)
     # Add some nodes in x and z for padding to avoid reflections
     pad = int(padding*max(shape))
-    decay = 20*pad
+    decay = float(20*pad)
     nx += 2*pad
     nz += pad
     # Pad the velocity as well
@@ -389,7 +392,7 @@ def elastic_sh(spacing, shape, svel, dens, deltat, iterations, sources,
     yield u_t[:-pad, pad:-pad]
     # Time steps
     for t in xrange(1, iterations):
-        fatiando.seis._cwavefd.step_elastic_sh(u_tp1, u_t, u_tm1, nx, nz,
+        timestepper.step_elastic_sh(u_tp1, u_t, u_tm1, nx, nz,
             deltat, dx, dz, svel_pad, pad, decay)
         # Set the boundary conditions
         u_tp1[1,:] = u_tp1[2,:]
@@ -446,10 +449,10 @@ def elastic_psv(spacing, shape, pvel, svel, dens, deltat, iterations, xsources,
 
     """
     nz, nx = shape
-    dz, dx = spacing
+    dz, dx = (float(i) for i in spacing)
     # Add some nodes in x and z for padding to avoid reflections
     pad = int(padding*max(shape))
-    decay = 20*pad
+    decay = float(20*pad)
     nx += 2*pad
     nz += pad
     # Pad the velocity as well
@@ -472,9 +475,9 @@ def elastic_psv(spacing, shape, pvel, svel, dens, deltat, iterations, xsources,
     yield ux_t[:-pad, pad:-pad], uz_t[:-pad, pad:-pad]
     # Time steps
     for t in xrange(1, iterations):
-        fatiando.seis._cwavefd.step_elastic_psv_x(ux_tp1, ux_t, ux_tm1, uz_t,
+        timestepper.step_elastic_psv_x(ux_tp1, ux_t, ux_tm1, uz_t,
             nx, nz, deltat, dx, dz, pvel_pad, svel_pad, pad, decay)
-        fatiando.seis._cwavefd.step_elastic_psv_z(uz_tp1, uz_t, uz_tm1, ux_t,
+        timestepper.step_elastic_psv_z(uz_tp1, uz_t, uz_tm1, ux_t,
             nx, nz, deltat, dx, dz, pvel_pad, svel_pad, pad, decay)
         # Set the boundary conditions
         ux_tp1[1,:] = ux_tp1[2,:]